backend/requirements.txt
fastapi
uvicorn
amadeus
boto3
python-dotenv